@@ -1,5 +1,6 @@
 {
   "name": "@tlecommunity/client",
+  "type": "module",
   "version": "0.3.0",
   "description": "API Client for the community edition of The Lacuna Expanse",
   "files": [
@@ -25,22 +26,22 @@
   "author": "Natalie McCallum <natalie@omg.lol>",
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^29.2.3",
-    "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
-    "prettier": "^2.7.1",
-    "ts-jest": "^29.0.3",
-    "typescript": "^4.9.3",
-    "vite": "^3.2.4"
+    "@types/jest": "^29.5.11",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "^3.1.1",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.10"
   },
   "engines": {
-    "node": ">=16.0.0",
-    "npm": ">=8.0.0"
+    "node": ">=20",
+    "npm": ">=8"
   },
   "dependencies": {
-    "@types/lodash": "^4.14.189",
-    "axios": "^1.1.3",
+    "@types/lodash": "^4.14.202",
+    "axios": "^1.6.3",
     "lodash": "^4.17.21",
-    "moment": "^2.29.4"
+    "moment": "^2.30.1"
   }
 }
