@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   // A list of paths to directories that Jest should use to search for files in
   // https://jestjs.io/docs/configuration#roots-arraystring
   roots: ['<rootDir>/lib/'],
